requirements.txt
openai-whisper
torch
tqdm
